requirements.txt
Django==6.0.2
djangorestframework==3.16.1
mysqlclient==2.2.8
cryptography==46.0.5
python-dotenv==1.2.2